feat: send mail registered

+++ app/Repositories/ProductRepository.php
@@ -4,18 +4,22 @@ namespace App\Repositories;
 
 use App\Interfaces\ProductRepositoryInterface;
 use App\Models\Product;
+use Cloudinary\Cloudinary;
 use Illuminate\Database\Eloquent\Model;
+use Illuminate\Http\UploadedFile;
+use Illuminate\Support\Facades\Storage;
 
 class ProductRepository implements ProductRepositoryInterface
 {
-    protected Product $product;
+    private Product $product;
 
     public function __construct(Product $product)
     {
         $this->product = $product;
     }
 
-    public function paginate($perPage = 10)
+
+    public function paginate($perPage = 500)
     {
         return $this->product->select()->orderBy('id', 'desc')->paginate($perPage);
     }
@@ -27,25 +31,34 @@ class ProductRepository implements ProductRepositoryInterface
 
     public function index()
     {
-        return $this->product->select('id', 'name', )->orderBy('id', 'desc')->get();
+        return $this->product->get();
     }
 
-    public function store(array $data)
+    public function create(array $data, )
     {
-        return $this->product->create($data);
+
+        return Product::create($data);
     }
 
-    public function update($id, array $data)
+    public function update($id, array $data,)
     {
         $product = Product::find($id);
-        $product->update($data);
-        return $product;
+
+            $product->update($data);
+            return $product;
+
     }
 
     public function delete($id)
     {
         $product = Product::find($id);
-        $product->delete();
+        if ($product) {
+            // Delete associated image
+            if ($product->image) {
+                Storage::disk('public')->delete($product->image);
+            }
+            return $product->delete();
+        }
         return $product;
     }
 }
